art & osc

+++ fin/art.h
@@ -11,10 +11,10 @@ class ART {
 public:
 	ART(unsigned long t_M, unsigned long t_N, float t_vigilance);
 	~ART();
-	long eval(float** x, unsigned long id);
-	unsigned long get_neuronCount();		
+	short eval(float** x, unsigned long id);
+	unsigned long get_neuronCount();
 	unsigned long get_id_of(unsigned long i);
-	
+
 private:
 	unsigned long M;
 	unsigned long N;
@@ -25,9 +25,9 @@ private:
 	float c;
 
 	void add_Neuron(float** x, unsigned long id);
-		
+
 	void save_State();
-		
+
 };
 
 #endif
